Fix some bugs
- Add quantity for merchandise
- Update OrdersDetail amount int to Long
- Update ReceiptDetail amount int to Long
- Add quantity limit in Orders (1000)
- Only show incomplete orders in add receipt
- Change ComboBox to auto complete textfield
- Update merchandise quantity when add new orders
- Add check valid email

+++ Software/src/main/java/controllers/ReceiptCategoryController.java
@@ -70,9 +70,9 @@ public class ReceiptCategoryController implements Initializable {
                 List<ReceiptModel> receiptModelList = new ArrayList<>();
                 for (Receipt item : receiptList) {
                     session = sessionFactory.openSession();
-                    Integer sumAmount = Math.toIntExact(OrdersDetailRepository.getSumAmountById(session, item.getOrders().getId()));
+                    Long sumAmount = OrdersDetailRepository.getSumAmountById(session, item.getOrders().getId());
                     session = sessionFactory.openSession();
-                    Integer sumQuantity = Math.toIntExact(OrdersDetailRepository.getSumQuantityById(session, item.getOrders().getId()));
+                    int sumQuantity = Math.toIntExact(OrdersDetailRepository.getSumQuantityById(session, item.getOrders().getId()));
 
                     ReceiptModel receiptModel = new ReceiptModel();
                     receiptModel.setReceipt(item);
